renamed jdbc package properly to the new 0.4 naming conventions
