+++ stratosphere-addons/jdbc/src/main/java/eu/stratosphere/api/java/record/io/jdbc/JDBCInputFormat.java
@@ -10,7 +10,7 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  **********************************************************************************************************************/
-package eu.stratosphere.api.io.jdbc;
+package eu.stratosphere.api.java.record.io.jdbc;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
